pyproject.toml
[project]
name = "pipo-objectstore"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.42.6",
    "fastapi>=0.124.4",
    "hdbcli>=2.26.26",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "sap-ai-sdk-gen>=5.9.0",
    "tabulate>=0.9.0",
    "uvicorn>=0.38.0",
]
